Extract plan from base.py into plan.py
